feat: CRUD de certificados de avance en proceso

+++ src/main/java/com/gestionobra/app/gestion_obra_app/dtos/CertificadoMaterialDTO.java
@@ -10,5 +10,5 @@ import lombok.Setter;
 @NoArgsConstructor
 public class CertificadoMaterialDTO {
     private Long materialId;
-    private Integer cantidadUtilizada;
+    private Double cantidadUtilizada;
 }
